feat: implement core controllers and services for orders, bills, conversations, messages, notifications, and statistics
